Change version lookup for packaging.
The current approach imports the version directly from crmngr. This
is a problem if the build environment does not have crmngr runtime
libraries installed and leads to having unnecessary build dependencies.
The new approach extracts the version from the python file directly,
allowing to have a single source of version, whitout needing to import
runtime dependencies during the build process.